MapF: Add a `BinTree.Internal` module, more `SPECIALIZE`
Fully removes dictionaries from the core.
